backend/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
python-dotenv==1.0.1
google-genai==1.66.0
pgvector==0.4.2
psycopg[binary]==3.2.6
gunicorn==23.0.0
psycopg2-binary==2.9.11
google-api-python-client==2.166.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
langchain-google-genai==2.1.3
